xml/parsing/manakai/entities-peref-1.dat

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x"/>
#resource http://a/
<!ENTITY % foo "">
<!ENTITY%foo;bar "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="aax"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x"/>
#resource http://a/
<!ENTITY % foo "">
<!ENTITY %foo; bar "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="aax"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x"/>
#resource http://a/
<!ENTITY % foo "">
<!ENTITY
   %foo; bar "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="aax"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x"/>
#resource http://a/
<!ENTITY % foo "  ">
<!ENTITY%foo;bar "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="aax"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x"/>
#resource http://a/
<!ENTITY % foo ">">
<!ENTITY%foo; bar "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;17;mdc in pe in md;;;m
[1]1;1;xml:dtd:ext decl;;;w
1;38;entity not declared;;&bar;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="&bar;x"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x"/>
#resource http://a/
<!ENTITY % foo ">">
<!ENTITY %foo; bar "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;17;mdc in pe in md;;;m
[1]1;1;xml:dtd:ext decl;;;w
1;38;entity not declared;;&bar;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="&bar;x"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x"/>
#resource http://a/
<!ENTITY % baz "">
<!ENTITY % foo "%baz;">
<!ENTITY %foo; bar "aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="aax"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x"/>
#resource http://a/
<!ENTITY % foo "bar">
<!ENTITY%foo;"aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="aax"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x"/>
#resource http://a/
<!ENTITY % foo "bar">
<!ENTITY  %foo;"aa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="aax"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x" c="&aa;"/>
#resource http://a/
<!ENTITY % foo "&#x25;">
<!ENTITY%foo;  bar "aa">
<!ENTITY %bar; "zz">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
1;38;entity not declared;;&bar;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="&bar;x"
|   c="zz"

#data
<!DOCTYPE a SYSTEM "http://a/"><a b="&bar;x" c="&aa;"/>
#resource http://a/
<!ENTITY % foo % bar "aaa">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
1;38;entity not declared;;&bar;;m
1;49;entity not declared;;&aa;;m
[1]1;17;no refc;;;m
[1]1;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   b="&bar;x"
|   c="&aa;"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&cd;"/>
#resource http://a/
<!ENTITY % fuga "">
<!ENTITY %fuga;cd "abc">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="abc"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&cd;"/>
#resource http://a/
<!ENTITY % fuga "">
<!ENTITY %fuga; cd "abc">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="abc"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&cd;"/>
#resource http://a/
<!ENTITY % fuga "cd">
<!ENTITY %fuga; "abc">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="abc"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&abc;"/>
#resource http://a/
<!ENTITY % fuga "&#x25; cd">
<!ENTITY %fuga; "abc">
<!ENTITY %cd; "zzz">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[1]3;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="zzz"

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge SYSTEM ""NDATA%abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[1]2;24;no space between params;;;m
[1]2;34;no notation name;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "xy">
<!ENTITY hoge SYSTEM ""NDATA%abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[1]2;24;no space between params;;;m
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "" xy>
| <{} a>
#entities
<!ENTITY hoge "" "" "" xy>
<!ENTITY % abc "xy" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "hoge SYSTEM 'x' NDATA">
<!ENTITY %abc;foo>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "x" foo>
| <{} a>
#entities
<!ENTITY hoge "" "" "x" foo>
<!ENTITY % abc "hoge SYSTEM 'x' NDATA" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "hoge SYSTEM 'x' NDATA>">
<!ENTITY %abc;foo>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;38;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "x" >
<!ENTITY % abc "hoge SYSTEM 'x' NDATA>" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "hoge SYSTEM 'x' NDATA ">
<!ENTITY %abc;foo>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "x" foo>
| <{} a>
#entities
<!ENTITY hoge "" "" "x" foo>
<!ENTITY % abc "hoge SYSTEM 'x' NDATA " "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge SYSTEM 'x' NDATA %abc;foo>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "x" foo>
| <{} a>
#entities
<!ENTITY hoge "" "" "x" foo>
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "hoge SYSTEM 'x' NDATA foo">
<!ENTITY %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "x" foo>
| <{} a>
#entities
<!ENTITY hoge "" "" "x" foo>
<!ENTITY % abc "hoge SYSTEM 'x' NDATA foo" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "foo">
<!ENTITY hoge SYSTEM 'x' NDATA %abc; >
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "x" foo>
| <{} a>
#entities
<!ENTITY hoge "" "" "x" foo>
<!ENTITY % abc "foo" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge SYSTEM 'x' NDATA foo%abc; >
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "x" foo>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge SYSTEM 'x' NDATA foo
%abc; >
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "x" foo>
| <{} a>
#entities
<!ENTITY hoge "" "" "x" foo>
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "fuga">
<!ENTITY hoge SYSTEM %abc; %xya;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;no DOCTYPE literal;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "" >
<!ENTITY % abc "fuga" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "fuga">
<!ENTITY hoge SYSTEM "b" %abc; %xya;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;string after SYSTEM literal;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "b" >
<!ENTITY % abc "fuga" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc ">a">
<!ENTITY hoge SYSTEM %abc; %xya;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "" >
<!ENTITY % abc ">a" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "fuga">
<!ENTITY hoge SYSTEM "" NDATA ab %abc; %xya;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;17;string after md def;;;m
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "" ab>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "fuga>">
<!ENTITY hoge SYSTEM "" NDATA ab %abc; %xya;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;string after md def;;;m
[1]1;21;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "" "" ab>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "fuga">
<!ENTITY%abc; "foo">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY fuga "foo" "" "" >
<!ENTITY % abc "fuga" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "fuga">
<!ENTITY %abc; "foo">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY fuga "foo" "" "" >
<!ENTITY % abc "fuga" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY fuga%abc; "foo">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY fuga "foo" "" "" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY fuga %abc; "foo">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY fuga "foo" "" "" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc ">">
<!ENTITY fuga%abc; "foo">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY fuga "" "" "" >
<!ENTITY % abc ">" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a/>
#resource http://a/
<!ENTITY % abc ">">
<!ENTITY fuga %abc; "foo">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY fuga "" "" "" >
<!ENTITY % abc ">" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge PUBLIC%abc;"a" "b">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "a" "b" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge PUBLIC %abc;"a" "b">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "a" "b" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "PUBLIC">
<!ENTITY hoge %abc;"a" "b">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "a" "b" >
<!ENTITY % abc "PUBLIC" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "PUBLIC ">
<!ENTITY hoge %abc;"a" "b">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "a" "b" >
<!ENTITY % abc "PUBLIC " "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "PUBLI">
<!ENTITY hoge %abc;C"a" "b">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;22;xml:dtd:string after name;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "" >
<!ENTITY % abc "PUBLI" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "PUBLIC >">
<!ENTITY hoge %abc;"a" "b">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;24;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "" >
<!ENTITY % abc "PUBLIC >" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc ">">
<!ENTITY hoge PUBLIC%abc;"a" "b">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "" >
<!ENTITY % abc ">" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc ">">
<!ENTITY hoge PUBLIC %abc;"a" "b">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "" >
<!ENTITY % abc ">" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc '"abc'>
<!ENTITY hoge PUBLIC %abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;unclosed literal;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "abc" "" >
<!ENTITY % abc ""abc" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "'abc">
<!ENTITY hoge PUBLIC %abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;unclosed literal;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "abc" "" >
<!ENTITY % abc "'abc" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc '"a>b"'>
<!ENTITY hoge PUBLIC %abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;19;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "a" "" >
<!ENTITY % abc ""a>b"" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "'a>b'">
<!ENTITY hoge PUBLIC %abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;19;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "a" "" >
<!ENTITY % abc "'a>b'" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "xx">
<!ENTITY hoge PUBLIC "%abc;" "a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "%abc;" "a" >
<!ENTITY % abc "xx" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "xx">
<!ENTITY hoge PUBLIC '%abc;' "a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "%abc;" "a" >
<!ENTITY % abc "xx" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge PUBLIC "b"%abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "a" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge PUBLIC "b" %abc; "a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "a" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc ">">
<!ENTITY hoge PUBLIC "b"%abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "" >
<!ENTITY % abc ">" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc ">">
<!ENTITY hoge PUBLIC "b" %abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;17;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "" >
<!ENTITY % abc ">" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge SYSTEM%abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "a" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge SYSTEM %abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "" "a" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge PUBLIC "b"%abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "a" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "">
<!ENTITY hoge PUBLIC "b" %abc;"a">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "a" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc '"ab'>
<!ENTITY hoge PUBLIC "b" %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;unclosed literal;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "ab" >
<!ENTITY % abc ""ab" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "'ab">
<!ENTITY hoge PUBLIC "b" %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;unclosed literal;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "ab" >
<!ENTITY % abc "'ab" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc '"a>b"'>
<!ENTITY hoge PUBLIC "b" %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "a>b" >
<!ENTITY % abc ""a>b"" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc '"a>b"'>
<!ENTITY hoge PUBLIC "b" %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "a>b" >
<!ENTITY % abc ""a>b"" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc ''>
<!ENTITY hoge PUBLIC "b" "%abc;">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "%abc;" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc ''>
<!ENTITY hoge PUBLIC "b" '%abc;'>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "%abc;" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc ''>
<!ENTITY hoge PUBLIC "b" ''%abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc ''>
<!ENTITY hoge PUBLIC "b" '' %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "" >
<!ENTITY % abc "" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc 'NDATA'>
<!ENTITY hoge PUBLIC "b" ''%abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[1]2;33;no notation name;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "" >
<!ENTITY % abc "NDATA" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc 'NDATA '>
<!ENTITY hoge PUBLIC "b" ''%abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[1]2;33;no notation name;;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "" "b" "" >
<!ENTITY % abc "NDATA " "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc 'NDATA'>
<!ENTITY hoge PUBLIC "b" ''%abc;abb>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ENTITY hoge "" "b" "" abb>
| <{} a>
#entities
<!ENTITY hoge "" "b" "" abb>
<!ENTITY % abc "NDATA" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc '"foo"'>
<!ENTITY hoge %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "foo" "" "" >
<!ENTITY % abc ""foo"" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "'foo'">
<!ENTITY hoge %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "foo" "" "" >
<!ENTITY % abc "'foo'" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc '"foo'>
<!ENTITY hoge %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;unclosed literal;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "foo" "" "" >
<!ENTITY % abc ""foo" "" "" >

#data
<!DOCTYPE a SYSTEM "http://a/" ><a/>
#resource http://a/
<!ENTITY % abc "'foo">
<!ENTITY hoge %abc;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;unclosed literal;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
#entities
<!ENTITY hoge "foo" "" "" >
<!ENTITY % abc "'foo" "" "" >

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f ">"><!ENTITY a SYSTEM "" NDATA x %f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;15;mdc in pe in md;;;m
[1]1;18;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
|   <!ENTITY a "" "" "" x>
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f ">"><!ENTITY a PUBLIC "a"%f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;15;mdc in pe in md;;;m
[1]1;18;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f ">"><!ENTITY a PUBLIC "a" %f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;15;mdc in pe in md;;;m
[1]1;18;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f '"a">'><!ENTITY a PUBLIC %f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;18;mdc in pe in md;;;m
[1]1;21;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f 'PUBLIC>'><!ENTITY a %f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;21;mdc in pe in md;;;m
[1]1;24;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f '"a">'><!ENTITY a SYSTEM %f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;18;mdc in pe in md;;;m
[1]1;21;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f 'SYSTEM>'><!ENTITY a %f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;21;mdc in pe in md;;;m
[1]1;24;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f '>'><!ENTITY%f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;15;mdc in pe in md;;;m
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f 'foo SYSTEM "x" NDATA>'><!ENTITY%f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;35;mdc in pe in md;;;m
[1]1;38;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f 'foo SYSTEM "x" NDATA ab>'><!ENTITY%f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;38;mdc in pe in md;;;m
[1]1;41;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
|   <!ENTITY foo "" "" "x" ab>
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f 'foo SYSTEM "x" NDATA >'><!ENTITY%f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;36;mdc in pe in md;;;m
[1]1;39;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>

#data
<?xml version="1.0"?><!DOCTYPE x SYSTEM "http://a/"><x/>
#resource http://a/
<!ENTITY % f 'foo SYSTEM "x"  >'><!ENTITY%f;>
#errors
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;31;mdc in pe in md;;;m
[1]1;34;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE x "" "http://a/">
| <{} x>
